+++ code/foot.h
@@ -37,6 +37,8 @@
 #include "team.h"
 #include "techno.h"
 
+#include <cstdint>
+
 class UnitClass;
 class BuildingClass;
 class WaypointClass;
@@ -389,7 +391,7 @@ class FootClass : public TechnoClass
 
 		virtual void Compute_CRC(CRCEngine &) const override;
 		virtual Coord Destination_Coord(void) const override;
-		virtual RadioMessageType Receive_Message(RadioClass * from, RadioMessageType message, int & param) override;
+		virtual RadioMessageType Receive_Message(RadioClass * from, RadioMessageType message, intptr_t & param) override;
 		virtual bool Can_Demolish(void) const override;
 		bool Is_Recruitable(HouseClass const * house=NULL) const;
 		bool Is_On_Priority_Mission(void) const;
